+++ CHANGELOG.md
@@ -13,6 +13,10 @@ This file contains the full list of changes for every version of Cherry Doom.
 
 - Reverted all rearrangements made to Nugget Doom's menu items and moved Cherry Doom's display options into a separate tab
 
+### Fixed
+
+- Dark backdrop fade-in/out animations not always working
+
 ## 2.1.0
 
 **Cherry Doom 2.1.0**, released June 30th, 2026. This version is up to date with [_Nugget Doom 5.1.0_](https://github.com/MrAlaux/Nugget-Doom/releases/tag/nugget-doom-5.1.0) and [_Woof! 15.3.0_](https://github.com/fabiangreffrath/woof/releases/tag/woof_15.3.0).